+++ lib/chewie.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'src/cast/index.dart';
 export 'src/chewie_player.dart';
 export 'src/chewie_progress_colors.dart';
 export 'src/cupertino/cupertino_controls.dart';